% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_hotspots.R
\name{drop_effort_cols}
\alias{drop_effort_cols}
\title{Drop effort columns from prob_per_loc}
\usage{
drop_effort_cols(prob_per_loc)
}
\arguments{
\item{prob_per_loc}{A tibble with the probability of observing each species at each locality}
}
\value{
A new copy of prob_per_loc without the effort columns
}
\description{
Drop effort columns from prob_per_loc
}
\examples{
drop_effort_cols(sample_prob_per_loc)
}
