% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_probability_matrix.R
\name{effort_per_hotspot}
\alias{effort_per_hotspot}
\title{Calculate effort information for each hotspot}
\usage{
effort_per_hotspot(observations)
}
\arguments{
\item{observations}{A tibble of eBird data}
}
\value{
A tibble with statistics on the sampling effort at each hotspot
}
\description{
Calculate effort information for each hotspot
}
