% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_probability_matrix.R
\name{n_observations_per_hotspot}
\alias{n_observations_per_hotspot}
\title{Calculate how many times each species was reported at each hotspot}
\usage{
n_observations_per_hotspot(observations)
}
\arguments{
\item{observations}{A tibble of eBird data}
}
\value{
A tibble with number of checklists reporting each species at each hotspot
}
\description{
Calculate how many times each species was reported at each hotspot
}
