% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_hotspots.R
\name{pred_total}
\alias{pred_total}
\title{Return the total expected number of species if visiting all suggested hotspots}
\usage{
pred_total(hotspots, probs)
}
\arguments{
\item{hotspots}{a character vector of hotspots of interest}

\item{probs}{A tibble with the probability of observing each species at each locality}
}
\value{
A numerical value - the predicted total of species at this set of hotspots
}
\description{
Return the total expected number of species if visiting all suggested hotspots
}
\examples{
hotspots <- c('Otter View Park', 'Hurd Grassland')
pred_total(hotspots, drop_effort_cols(sample_prob_per_loc))
}
