% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hotspots.R
\name{prob_hotspots}
\alias{prob_hotspots}
\title{Calculate the probability of seeing each species for a given set of hotspots}
\usage{
prob_hotspots(hotspots, prob_per_loc)
}
\arguments{
\item{hotspots}{A character vector of hotspot names}

\item{prob_per_loc}{A tibble with the probability of observing each species at each locality}
}
\value{
A named numerical vector of the probability of observing each species at the given combination of hotspots
}
\description{
Calculate the probability of seeing each species for a given set of hotspots
}
\examples{
# hotspots <- c('Otter View Park', 'Button Bay State Park')
# prob_hotspots(hotspots, prob_per_loc)
}
