% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_probability_matrix.R
\name{probability_matrix}
\alias{probability_matrix}
\title{Calculate probability of seeing each species at each hotspot}
\usage{
probability_matrix(observations)
}
\arguments{
\item{observations}{A tibble of eBird data}
}
\value{
A matrix with probability of seeing each species at each hotspot
}
\description{
Calculate probability of seeing each species at each hotspot
}
