% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_counties}
\alias{sample_counties}
\title{sample county list, containing just Addison County}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
sample_counties
}
\description{
sample county list, containing just Addison County
}
\keyword{datasets}
