% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_hotspots}
\alias{sample_hotspots}
\title{sample list of hotspots in Addison County}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 5 columns.
}
\usage{
sample_hotspots
}
\description{
sample list of hotspots in Addison County
}
\keyword{datasets}
