% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_observations}
\alias{sample_observations}
\title{sample eBird observation data for Addison County, VT, US, during April 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21266 rows and 20 columns.
}
\source{
\url{https://ebird.org/data/download/ebd}
}
\usage{
sample_observations
}
\description{
sample eBird observation data for Addison County, VT, US, during April 2020
}
\keyword{datasets}
