% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_prob_per_loc}
\alias{sample_prob_per_loc}
\title{sample tibble of probability of seeing each species at each hotspot
in Addison County, calculated from sample_observations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57 rows and 205 columns.
}
\usage{
sample_prob_per_loc
}
\description{
sample tibble of probability of seeing each species at each hotspot
in Addison County, calculated from sample_observations
}
\keyword{datasets}
