% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_find_hotspots.R
\name{select_hotspots}
\alias{select_hotspots}
\title{Greedy selection algorithm to select the k optimal hotspots to visit}
\usage{
select_hotspots(probs, k, H = NULL)
}
\arguments{
\item{probs}{A tibble with the probability of observing each species at each locality}

\item{k}{The number of hotspots you want to visit}

\item{H}{A character vector of the hotspots you definitely want to visit}
}
\description{
Greedy selection algorithm to select the k optimal hotspots to visit
}
\examples{
# Not run because of long runtime
# hotspots <- c('Otter View Park', 'Button Bay State Park')
# select_hotspots(drop_effort_cols(prob_per_loc), 5, hotspots)
}
