% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_probability_matrix.R
\name{write_prob_per_loc}
\alias{write_prob_per_loc}
\title{Save a single prob_per_loc matrix for a specified county}
\usage{
write_prob_per_loc(observations, loc)
}
\arguments{
\item{observations}{A tibble of eBird data}

\item{loc}{A string denoting the county code}
}
\value{
invisibly returns the prob_per_loc matrix
}
\description{
Save a single prob_per_loc matrix for a specified county
}
