% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_probability_matrix.R
\name{write_probs_all_counties}
\alias{write_probs_all_counties}
\title{Save a prob_per_loc matrix for every county in the dataset}
\usage{
write_probs_all_counties(observations)
}
\arguments{
\item{observations}{A tibble of eBird data}
}
\value{
invisibly returns the last prob_per_loc matrix written
TODO this return behavior is kind of funky...
}
\description{
Save a prob_per_loc matrix for every county in the dataset
}
